+++ _universum/submit.py
@@ -19,6 +19,8 @@ class Submit(Module):
 
     @staticmethod
     def define_arguments(parser):
+        parser.set_defaults(main_class=Submit)
+
         parser.add_argument("--create-review", action="store_true", dest="review",
                             help="create deletable review (shelve for P4, temp branch for Git) "
                                  "instead of actual submitting to repo")